fix torch call

+++ models_torch.py
@@ -465,7 +465,7 @@ class Agent:
 
     def act(self, environment):
         state  = environment.get_state()
-        action = self.get_action(torch.from_numpy(state))
+        action = self.get_action(torch.from_numpy(state, dtype=torch.double))
         next_state, reward = environment.perform_action(action)
         self.store_transition((state, action, reward, next_state))
         return reward
